requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.4
psycopg2-binary==2.9.10
python-decouple==3.8
django-filter==25.1
Pillow==10.2.0
django-widget-tweaks==1.5.0
django-tables2==2.7.0
django-filter==25.1
django-import-export==3.3.6
django-cors-headers==4.3.1 